backend/requirements.txt
flask
flask-cors
gunicorn
pandas
apscheduler
unidecode
numpy
scipy
opentelemetry-api
opentelemetry-sdk
opentelemetry-ext-otcollector